Prototype complete
The basic prototype of this app is complete. This commit is a functional
application which implements **extralive chat**, that is, it reflects
keystrokes and characters directly to the end users as they are typed.
The app does not have any other features. All messages are anonymous,
messages are stored in RAM with no db, and there is no moderation or
voting.

This app just contains a client and a server that implement extralive
chat. The client and server both have a single monolithic main.rs file.
Now that a functional prototype / proof of concept exists, the next step
in the development of this program is a massive refactor.